move the protos to pkg/protos
