+++ .devcontainer/go/postCreateCommand.sh
@@ -9,7 +9,7 @@ make protos-all
 
 cd ${ROOT_DIR}
 rm -rf ${ROOT_DIR}/go.work ${ROOT_DIR}/go.work.sum
-go work init ./sdks/go ./build/go
+go work init ./sdks/go
 go work edit -go=1.24
 go work edit -toolchain=go1.24.11
 go work sync
