+++ .vscode/launch.json
@@ -7,6 +7,7 @@
         "request": "launch",
         "mode": "debug",
         "program": "${workspaceFolder}/sdks/go/examples/${input:pickGoService}",
+        "preLaunchTask": "Generate Protobufs",
         "cwd": "${workspaceFolder}",
         "args": [],
         "env": {
