+++ sdks/go/go.mod
@@ -5,7 +5,6 @@ go 1.24.0
 toolchain go1.24.13
 
 require (
-	github.com/rossvideo/catena/build/go v0.0.0
 	google.golang.org/grpc v1.79.1
 	google.golang.org/protobuf v1.36.11
 )
@@ -16,5 +15,3 @@ require (
 	golang.org/x/text v0.32.0 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20251202230838-ff82c1b0f217 // indirect
 )
-
-replace github.com/rossvideo/catena/build/go => ../../build/go
