+++ sdks/go/pkg/catena/command_test.go
@@ -41,7 +41,7 @@ package catena
 import (
 	"testing"
 
-	"github.com/rossvideo/catena/build/go/protos"
+	"github.com/rossvideo/catena/sdks/go/pkg/protos"
 )
 
 func TestNewPolyglotText(t *testing.T) {
