+++ sdks/go/pkg/catena/value.go
@@ -42,7 +42,7 @@ package catena
 import (
 	"fmt"
 
-	"github.com/rossvideo/catena/build/go/protos"
+	"github.com/rossvideo/catena/sdks/go/pkg/protos"
 )
 
 type CatenaValue struct {
