+++ sdks/go/pkg/internal/connection_queue_test.go
@@ -34,7 +34,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/rossvideo/catena/build/go/protos"
+	"github.com/rossvideo/catena/sdks/go/pkg/protos"
 )
 
 func TestNewConnectionQueue(t *testing.T) {
